% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{bar-class}
\alias{bar-class}
\alias{bar}
\title{Bar}
\value{
a class of type "bar"
}
\description{
Bar
}
\section{Slots}{

\describe{
\item{\code{distance}}{numeric.}
}}

\examples{
bar <- new("bar", x = pi, y = 2, distance = 0)
bar["x"]
bar["distance"]
}
