% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{[,bar-method}
\alias{[,bar-method}
\title{Bar methods}
\usage{
\S4method{[}{bar}(x, i, j, ..., bar = TRUE, drop = FALSE)
}
\arguments{
\item{x}{bar.}

\item{i}{ANY.}

\item{j}{ANY.}

\item{...}{passed on to foo}

\item{bar}{When TRUE, it prints a message}

\item{drop}{ANY.}
}
\value{
an object of class "bar"
}
\description{
Bar methods
}
\examples{
bar <- new("bar", x = pi, y = 2, distance = 0)
bar["x"]
bar["distance"]
}
