% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{foo-class}
\alias{foo-class}
\alias{foo}
\title{Foo}
\value{
a class of type "foo"
}
\description{
Foo
}
\section{Slots}{

\describe{
\item{\code{x}}{numeric.}

\item{\code{y}}{numeric.}
}}

\examples{
foo <- new("foo", x = pi, y = 2)
foo["x"]
foo["y"]
}
