% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{[,foo-method}
\alias{[,foo-method}
\title{Foo methods}
\usage{
\S4method{[}{foo}(x, i, j, ..., foo = TRUE, drop = FALSE)
}
\arguments{
\item{x}{foo.}

\item{i}{ANY.}

\item{j}{ANY.}

\item{...}{not used}

\item{foo}{When TRUE, it prints a message}

\item{drop}{ANY.}
}
\value{
an object of class foo
}
\description{
Foo methods
}
\examples{
foo <- new("foo", x = pi, y = 2)
foo["x"]
}
