% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{initialize,bar-method}
\alias{initialize,bar-method}
\title{Initialize Bar}
\usage{
\S4method{initialize}{bar}(.Object, ..., distance)
}
\arguments{
\item{.Object}{a character, "bar"}

\item{...}{parameters passed on to foo}

\item{distance}{numeric}
}
\description{
Initialize Bar
}
