% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\docType{methods}
\name{initialize,foo-method}
\alias{initialize,foo-method}
\title{Initialize Foo}
\usage{
\S4method{initialize}{foo}(.Object, x, y)
}
\arguments{
\item{.Object}{a character, "foo"}

\item{x}{numeric}

\item{y}{numeric}
}
\description{
Initialize Foo
}
